Issue #94 - Move Authentication service and add isAdmin method
