+++ frontend/src/app/modules/authentication/authentication.service.spec.ts
@@ -15,27 +15,9 @@
  *  * limitations under the License.
  *
  */
-
-/*
- *  /*
- *  * Copyright 2012-2018 the original author or authors.
- *  *
- *  * Licensed under the Apache License, Version 2.0 (the "License");
- *  * you may not use this file except in compliance with the License.
- *  * You may obtain a copy of the License at
- *  *
- *  *      http://www.apache.org/licenses/LICENSE-2.0
- *  *
- *  * Unless required by applicable law or agreed to in writing, software
- *  * distributed under the License is distributed on an "AS IS" BASIS,
- *  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  * See the License for the specific language governing permissions and
- *  * limitations under the License.
- *
- */
 import {inject, TestBed} from '@angular/core/testing';
 
-import {AuthenticationService} from '../../modules/authentication/authentication.service';
+import {AuthenticationService} from './authentication.service';
 
 describe('AuthenticationService', () => {
   beforeEach(() => {
