+++ frontend/src/app/modules/user/user.module.ts
@@ -24,10 +24,11 @@ import {CommonModule} from '@angular/common';
 import {SharedModule} from '../../shared/shared.module';
 import {DeleteUserDialogComponent} from './components/delete-user-dialog/delete-user-dialog.component';
 import {UserEditComponent} from './user-edit/user-edit.component';
+import {AdminGuard} from '../../shared/auth/guards/admin.guard';
 
 const appRoutes: Routes = [
   {path: 'users', component: UserListComponent, data: {breadcrumb: 'User List'}, canActivate: [AuthGuard]},
-  {path: 'users/:userId/edit', component: UserEditComponent, data: {breadcrumb: 'Edit User'}, canActivate: [AuthGuard]}
+  {path: 'users/:userId/edit', component: UserEditComponent, data: {breadcrumb: 'Edit User'}, canActivate: [AuthGuard, AdminGuard]}
 ];
 
 @NgModule({
